requirements.txt
pandas
numpy
scikit-learn
nltk
streamlit
matplotlib
joblib
openai
python-dotenv
